added brackets

+++ main.cpp
@@ -103,7 +103,8 @@ static cmd_err_t read_serial_command(cmd_t * cmd) {
     char check = buf[3];
     char end = buf[4];
 
-    if(start != 2 && end != 3 && check != (buf[1] + buf[2])){
+    if(start != 2 && end != 3 && 
+            check != ((uint8_t)(((uint8_t)buf[1]) + ((uint8_t)buf[2])))){
         return CMD_ERR_INVALID;
     }
